cache_server/app/search/requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
uhashring==2.3
numpy==1.21.5
pandas==1.0.5
matplotlib==3.4.3
